plot typeの削除と確認画面の修正
